modified scripts to allow resizing, resized Scout
