+++ editions/free/src/localizations/en.json
@@ -237,8 +237,8 @@
     "SAMPLE_TEXT_key_bump_green_duck": "Green Duck",
     "SAMPLE_TEXT_key_quickintro_fish": "Fish",
     "SAMPLE_TEXT_key_quickintro_toc": "Toc",
-    "CHARACTER_Scout-walk-right-1.svg": "Scout",
-    "CHARACTER_Scout-point-right.svg": "Scout point",
+    "CHARACTER_Scout.svg": "Scout",
+    "CHARACTER_Scout-point.svg": "Scout point",
     "CHARACTER_Cat.svg": "Cat",
     "CHARACTER_Cat2.svg": "Cat Walking",
     "CHARACTER_Blue.svg": "Tic",
