+++ editions/free/src/localizations/es.json
@@ -179,8 +179,8 @@
     "SAMPLE_TEXT_key_bump_green_duck": "Pato Verde",
     "SAMPLE_TEXT_key_quickintro_fish": "Pez",
     "SAMPLE_TEXT_key_quickintro_toc": "Toc",
-    "CHARACTER_Scout-walk-right-1.svg": "Scout",
-    "CHARACTER_Scout-point-right.svg": "Scout apuntando",
+    "CHARACTER_Scout.svg": "Scout",
+    "CHARACTER_Scout-point.svg": "Scout apuntando",
     "CHARACTER_Cat.svg": "Gato",
     "CHARACTER_Cat2.svg": "Gato andando",
     "CHARACTER_Blue.svg": "Ta",
